+++ src/components/Card/CardList/CardList.js
@@ -1,7 +1,5 @@
 import "./CardList.css";
 
-const CardList = ({ children }) => {
+export const CardList = ({ children }) => {
   return <div className="cardList">{children}</div>;
 };
-
-export default CardList;
