+++ src/components/Card/CardWrapper/CardWrapper.js
@@ -1,11 +1,9 @@
 import "./CardWrapper.css";
 
-const Card = ({ children, onMouseOver, onMouseLeave }) => {
+export const CardWrapper = ({ children, onMouseOver, onMouseLeave }) => {
   return (
     <div className="card" onMouseOver={onMouseOver} onMouseLeave={onMouseLeave}>
       {children}
     </div>
   );
 };
-
-export default Card;
