+++ src/components/Card/OnlyCard/OnlyCard.js
@@ -1,9 +1,9 @@
 import { useState } from "react";
-import { CardContent } from "../CardContent/CardContent";
-import CardImage from "../CardImage/CardImage";
-import Card from "../CardWrapper/CardWrapper";
+import { CardImage } from "../CardImage";
+import { CardWrapper } from "../CardWrapper";
+import { CardContent } from "../CardContent";
 
-const OnlyCard = ({ items }) => {
+export const OnlyCard = ({ items }) => {
   const [isHovered, setIsHovered] = useState(false);
   const handleMouseOver = () => setIsHovered(true);
   const handleMouseLeave = () => setIsHovered(false);
@@ -14,12 +14,13 @@ const OnlyCard = ({ items }) => {
   };
   return (
     <a href={items.url} target="_blank" rel="noopener noreferrer">
-      <Card onMouseOver={handleMouseOver} onMouseLeave={handleMouseLeave}>
+      <CardWrapper
+        onMouseOver={handleMouseOver}
+        onMouseLeave={handleMouseLeave}
+      >
         <CardImage items={items} isZoomedIn={isHovered} />
         <CardContent items={items} handleLinkClick={handleLinkClick} />
-      </Card>
+      </CardWrapper>
     </a>
   );
 };
-
-export default OnlyCard;
